+++ pages/admin.js
@@ -1,0 +1,9 @@
+const admin = () => {
+    return (
+        <div>
+            <h1>ADMIN</h1>
+        </div>
+    )
+}
+
+export default admin
