+++ pages/api/user.js
@@ -17,3 +17,9 @@ export default async function handler(req, res) {
       );
   }
 }
+
+export const config = {
+  api: {
+    externalResolver: true,
+  },
+}
